+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix many QSV pipeline bugs
 - Fix MPEG2 video format with QSV and VAAPI acceleration
 - Fix playback of content with undefined colorspace
+- Fix NVIDIA color normalization with VP9 sources
 
 ## [0.7.3-beta] - 2023-01-25
 ### Added